Lesson 26; HW26

+++ WeatherZhuravlevMax/WeatherZhuravlevMax/Struct/DailyWeatherData.swift
@@ -18,7 +18,7 @@ struct DailyWeatherData: Codable {
     var windGust: Double?
     var weather: [WeatherDaily]?
     var clouds: Int?
-    var pop: Int?
+    var pop: Double?
     var rain: Double?
     var uvi: Double?
     
@@ -56,3 +56,4 @@ struct WeatherDaily: Codable {
     var description: String?
     var icon: String?
 }
+
